Fix content-type in GET requests
The content-type header should not be provided during a GET request as the request does not have a payload.
This behaviour breaks communication with the SignHost API, although there's currently a workaround in place. But the plan is to remove this workaround in the future, as sending the content-type header is not conform spec.